+++ lib/fpdart.dart
@@ -1,3 +1,4 @@
+export 'src/boolean.dart';
 export 'src/compose.dart';
 export 'src/either.dart';
 export 'src/function.dart';
@@ -5,6 +6,7 @@ export 'src/io.dart';
 export 'src/io_either.dart';
 export 'src/list_extension.dart';
 export 'src/option.dart';
+export 'src/random.dart';
 export 'src/reader.dart';
 export 'src/state.dart';
 export 'src/task.dart';